[update] : common eslint 적용
