+++ nbc-nep-front/src/components/common/SpaceThumb.tsx
@@ -7,6 +7,13 @@ interface Props {
   src: string | undefined;
 }
 
+const StImage = styled(Image)`
+  border-radius: ${(props) => props.theme.border.radius[8]};
+  padding: 0;
+  margin: 0;
+  object-fit: cover;
+`
+
 export default function SpaceThumb({
   width = 250,
   height = 150,
@@ -23,9 +30,3 @@ export default function SpaceThumb({
   );
 }
 
-const StImage = styled(Image)`
-  border-radius: ${(props) => props.theme.border.radius[8]};
-  padding: 0;
-  margin: 0;
-  object-fit: cover;
-`;
