+++ nbc-nep-front/src/components/common/badge/BadgeNumber.tsx
@@ -1,6 +1,5 @@
-import { Interactive, Interactive2 } from "@/types/theme.types";
-import styled from "styled-components";
-import StBadge from "./Badge";
+import { Interactive, Interactive2 } from '@/types/theme.types';
+import { StBadgeNumber } from './badge.styles';
 
 interface Props {
   count: number;
@@ -8,6 +7,9 @@ interface Props {
   backgroundColor?: keyof Interactive;
   color?: keyof Interactive2;
 }
+
+
+
 export default function BadgeNumber({
   count,
   maxCount = 10,
@@ -20,28 +22,7 @@ export default function BadgeNumber({
     <StBadgeNumber $bg={backgroundColor} $color={color}>
       <span>{displayCount}</span>
     </StBadgeNumber>
-  );
+  )
 }
 
-const StBadgeNumber = styled(StBadge)<{
-  $bg: keyof Interactive;
-  $color: keyof Interactive2;
-}>`
-  top: ${(props) => `calc(-1 * ${props.theme.spacing[48]})`};
-  left: ${(props) => props.theme.spacing[4]};
 
-  width: 20px;
-  height: 20px;
-  font-size: ${(props) => props.theme.body.md.medium.fontSize};
-  border-radius: ${(props) => props.theme.border.radius.circle};
-
-  display: flex;
-  justify-content: center;
-  align-items: end;
-  font-weight: bold;
-  background-color: ${(props) => props.theme.color.bg.interactive[props.$bg]};
-  color: ${(props) => props.theme.color.text.interactive[props.$color]};
-  > * {
-    color: ${(props) => props.theme.color.text.interactive[props.$color]};
-  }
-`;
