+++ nbc-nep-front/src/components/common/badge/BadgeWrapper.tsx
@@ -1,12 +1,9 @@
-import { PropsWithChildren } from "react";
-import styled from "styled-components";
+import { PropsWithChildren } from 'react';
+import { StBadgeWrapper } from './badge.styles';
+
+
 
 export default function BadgeWrapper({ children }: PropsWithChildren) {
   return <StBadgeWrapper>{children}</StBadgeWrapper>;
 }
 
-const StBadgeWrapper = styled.div`
-  position: relative;
-  width: 0;
-  height: 0;
-`;
