+++ nbc-nep-front/src/components/metaverse/metaversePlayerList/MetaversePlayerCard.tsx
@@ -1,36 +1,8 @@
-import styledComponents from "styled-components";
-
-import StBadge from "@/components/common/badge/Badge";
-import MetaAvatar from "@/components/metaverse/avatar/MetaAvatar";
-import { getPlayerStateValue } from "@/components/video-conference/DockPlayer";
-import useAuthStore from "@/zustand/authStore";
-import { HandleOpenDmContainerPrams, Player } from "@/types/metaverse.types";
-
-const styled = styledComponents;
-
-const StMetaversePlayerCard = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  height: 30px;
-  cursor: pointer;
-  > div {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-  }
-  > div > span {
-    font-size: 14px;
-    white-space: nowrap;
-  }
-  &:hover {
-    box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.5);
-  }
-`;
-
-const StBadgeWrapper = styled.div`
-  position: relative;
-`;
+import StBadge from '@/components/common/badge/badge.styles'
+import MetaAvatar from '@/components/metaverse/avatar/MetaAvatar'
+import { getPlayerStateValue } from '@/components/video-conference/DockPlayer'
+import { HandleOpenDmContainerPrams, Player } from '@/types/metaverse.types'
+import useAuthStore from '@/zustand/authStore'
 
 interface Props {
   player: Player;
