+++ nbc-nep-front/src/components/video-conference/DockPlayer.tsx
@@ -3,18 +3,13 @@ import { useState } from "react";
 import useAnimated from "@/hooks/useAnimated";
 import { Player } from "@/types/metaverse.types";
 
-import StBadge from "../common/badge/Badge";
-import BadgeWrapper from "../common/badge/BadgeWrapper";
-import MetaAvatar from "../metaverse/avatar/MetaAvatar";
+import StBadge from '../common/badge/badge.styles';
+import BadgeWrapper from '../common/badge/BadgeWrapper';
+import MetaAvatar from '../metaverse/avatar/MetaAvatar';
 
-import PlayerStateSelector from "./PlayerStateSelector";
-import { getPlayerStateToText, getPlayerStateValue } from "./libs/dock";
-import {
-  StDockPlayerInfoWrapper,
-  StDockPlayerNickname,
-  StDockPlayerState,
-  StDockPlayerWrapper,
-} from "./styles/dockPlayer.styles";
+import { getPlayerStateToText, getPlayerStateValue } from './libs/dock';
+import PlayerStateSelector from './PlayerStateSelector';
+import { StDockPlayerInfoWrapper, StDockPlayerNickname, StDockPlayerState, StDockPlayerWrapper } from './styles/dockPlayer.styles';
 
 interface Props {
   player?: Player;
